pyproject.toml
[project]
name = "uv-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.4.0",
    "redis[hiredis]>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.40",
]
